reorder functions and files
