+++ include/calculate.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-using Byte = unsigned char;
+#include "types.hpp"
 
 void* allocateGraphicsMemory();
 void freeGraphicsMemory(void* cudaPixels);
 void juliaFatouCUDA(Byte* textureImg, void* cudaPixels);
-void screenshot(Byte* pixels);
+void singleBigFrame(Byte* pixels);
