+++ include/types.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+using Byte = unsigned char;
